+++ test/utils.ts
@@ -9,14 +9,19 @@ export function loadAllLists() {
   return [
     'assets/easylist-downloads.adblockplus.org/antiadblockfilters.txt',
     'assets/easylist.to/easylist/easylist.txt',
+    'assets/easylist.to/easylist/easyprivacy.txt',
     'assets/easylist.to/easylistgermany/easylistgermany.txt',
     'assets/raw.githubusercontent.com/uBlockOrigin/uAssets/master/filters/badware.txt',
     'assets/raw.githubusercontent.com/uBlockOrigin/uAssets/master/filters/filters.txt',
     'assets/raw.githubusercontent.com/uBlockOrigin/uAssets/master/filters/resource-abuse.txt',
     'assets/raw.githubusercontent.com/uBlockOrigin/uAssets/master/filters/unbreak.txt',
-  ].map(readAsset).join('\n');
+  ]
+    .map(readAsset)
+    .join('\n');
 }
 
 export function loadResources() {
-  return readAsset('assets/raw.githubusercontent.com/uBlockOrigin/uAssets/master/filters/resources.txt');
+  return readAsset(
+    'assets/raw.githubusercontent.com/uBlockOrigin/uAssets/master/filters/resources.txt',
+  );
 }